+++ src/OpenTelemetry.Exporter.Zipkin/ZipkinExporterOptions.cs
@@ -16,6 +16,8 @@
 
 using System;
 using System.Diagnostics;
+using System.Linq;
+using OpenTelemetry.Resources;
 
 namespace OpenTelemetry.Exporter
 {
@@ -24,17 +26,9 @@ namespace OpenTelemetry.Exporter
     /// </summary>
     public sealed class ZipkinExporterOptions
     {
-        internal const string DefaultServiceName = "OpenTelemetry Exporter";
-
 #if !NET452
         internal const int DefaultMaxPayloadSizeInBytes = 4096;
 #endif
-
-        /// <summary>
-        /// Gets or sets the name of the service reporting telemetry. If the `Resource` associated with the telemetry
-        /// has "service.name" defined, then it'll be preferred over this option.
-        /// </summary>
-        public string ServiceName { get; set; } = DefaultServiceName;
 
         /// <summary>
         /// Gets or sets Zipkin endpoint address. See https://zipkin.io/zipkin-api/#/default/post_spans.